cleaning and change package name
