+++ app/src/main/java/cash/leeap/Data.java
@@ -1,0 +1,8 @@
+package cash.leeap;
+
+class Data {
+
+    static final String API_URL = "https://api.leeap.cash/";
+
+
+}
