+++ app/src/main/java/cash/leeap/Splash.java
@@ -1,0 +1,18 @@
+package cash.leeap;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+public class Splash extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        Intent intent = new Intent(this, Main.class);
+        startActivity(intent);
+        finish();
+    }
+}
+
