+++ app/src/main/java/cash/leeap/Store.java
@@ -1,0 +1,16 @@
+package cash.leeap;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.r0adkll.slidr.Slidr;
+
+public class Store extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_store);
+        Slidr.attach(this);
+    }
+}
