--- app/src/main/java/enib/gala/ScanBraceletActivity.java
@@ -1,203 +1,0 @@
-package enib.gala;
-
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
-import android.net.Uri;
-import android.nfc.NdefMessage;
-import android.nfc.NdefRecord;
-import android.nfc.NfcAdapter;
-import android.nfc.Tag;
-import android.os.Bundle;
-import android.os.Parcelable;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
-
-import com.r0adkll.slidr.Slidr;
-import com.r0adkll.slidr.model.SlidrConfig;
-import com.r0adkll.slidr.model.SlidrPosition;
-
-import java.io.UnsupportedEncodingException;
-
-public class ScanBraceletActivity extends AppCompatActivity { //https://github.com/codexpedia/android_nfc_read_write/
-
-    private String braceletValue;
-    private NfcAdapter mNfcAdapter;
-    boolean writeMode;
-    PendingIntent pendingIntent;
-    IntentFilter writeTagFilters[];
-    Tag myTag;
-    Context context;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_scan_bracelet);
-
-        SlidrConfig config = new SlidrConfig.Builder()
-                                .position(SlidrPosition.BOTTOM)
-
-                                .build();
-
-        Slidr.attach(this, config);
-
-        context = this;
-
-        mNfcAdapter = NfcAdapter.getDefaultAdapter(this);
-
-        if (mNfcAdapter == null) {
-            // Stop here, we definitely need NFC
-            Toast.makeText(this, "This device doesn't support NFC.", Toast.LENGTH_LONG).show();
-            finish();
-            return;
-        }
-
-        if (!mNfcAdapter.isEnabled()) {
-            Toast.makeText(this, "NFC is disabled.", Toast.LENGTH_LONG).show();
-        }
-
-        readFromIntent(getIntent());
-
-        pendingIntent = PendingIntent.getActivity(this, 0, new Intent(this, getClass()).addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP), 0);
-        IntentFilter tagDetected = new IntentFilter(NfcAdapter.ACTION_TAG_DISCOVERED);
-        tagDetected.addCategory(Intent.CATEGORY_DEFAULT);
-        writeTagFilters = new IntentFilter[] { tagDetected };
-
-        Button mButton= findViewById(R.id.button_scan_bracelet_action);
-        mButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
-    }
-
-    private void returnData()
-    {
-        Intent data = new Intent();
-        data.setData(Uri.parse(braceletValue));
-        setResult(RESULT_OK, data);
-        finish();
-    }
-
-
-    /******************************************************************************
-     **********************************Read From NFC Tag***************************
-     ******************************************************************************/
-    private void readFromIntent(Intent intent) {
-        String action = intent.getAction();
-        if (NfcAdapter.ACTION_TAG_DISCOVERED.equals(action)
-                || NfcAdapter.ACTION_TECH_DISCOVERED.equals(action)
-                || NfcAdapter.ACTION_NDEF_DISCOVERED.equals(action)) {
-            Parcelable[] rawMsgs = intent.getParcelableArrayExtra(NfcAdapter.EXTRA_NDEF_MESSAGES);
-            NdefMessage[] msgs = null;
-            if (rawMsgs != null) {
-                msgs = new NdefMessage[rawMsgs.length];
-                for (int i = 0; i < rawMsgs.length; i++) {
-                    msgs[i] = (NdefMessage) rawMsgs[i];
-                }
-            }
-            buildTagViews(msgs);
-        }
-    }
-    private void buildTagViews(NdefMessage[] msgs) {
-        if (msgs == null || msgs.length == 0) return;
-
-        String text = "";
-//        String tagId = new String(msgs[0].getRecords()[0].getType());
-        byte[] payload = msgs[0].getRecords()[0].getPayload();
-        String textEncoding = ((payload[0] & 128) == 0) ? "UTF-8" : "UTF-16"; // Get the Text Encoding
-        int languageCodeLength = payload[0] & 0063; // Get the Language Code, e.g. "en"
-        // String languageCode = new String(payload, 1, languageCodeLength, "US-ASCII");
-
-        try {
-            // Get the Text
-            text = new String(payload, languageCodeLength + 1, payload.length - languageCodeLength - 1, textEncoding);
-            braceletValue=text;
-            returnData();
-//            Toast.makeText(this, "NFC Content: " + text, Toast.LENGTH_LONG).show();
-        } catch (UnsupportedEncodingException e) {
-            Log.e("UnsupportedEncoding", e.toString());
-        }
-
-
-    }
-
-
-    /******************************************************************************
-     **********************************Write to NFC Tag****************************
-     ******************************************************************************/
-//    private void write(String text, Tag tag) throws IOException, FormatException {
-//        NdefRecord[] records = { createRecord(text) };
-//        NdefMessage message = new NdefMessage(records);
-//        // Get an instance of Ndef for the tag.
-//        Ndef ndef = Ndef.get(tag);
-//        // Enable I/O
-//        ndef.connect();
-//        // Write the message
-//        ndef.writeNdefMessage(message);
-//        // Close the connection
-//        ndef.close();
-//    }
-    private NdefRecord createRecord(String text) throws UnsupportedEncodingException {
-        String lang       = "en";
-        byte[] textBytes  = text.getBytes();
-        byte[] langBytes  = lang.getBytes("US-ASCII");
-        int    langLength = langBytes.length;
-        int    textLength = textBytes.length;
-        byte[] payload    = new byte[1 + langLength + textLength];
-
-        // set status byte (see NDEF spec for actual bits)
-        payload[0] = (byte) langLength;
-
-        // copy langbytes and textbytes into payload
-        System.arraycopy(langBytes, 0, payload, 1,              langLength);
-        System.arraycopy(textBytes, 0, payload, 1 + langLength, textLength);
-
-        return new NdefRecord(NdefRecord.TNF_WELL_KNOWN, NdefRecord.RTD_TEXT, new byte[0], payload);
-    }
-
-
-
-    @Override
-    protected void onNewIntent(Intent intent) {
-        setIntent(intent);
-        readFromIntent(intent);
-        if (NfcAdapter.ACTION_TAG_DISCOVERED.equals(intent.getAction())){
-            myTag = intent.getParcelableExtra(NfcAdapter.EXTRA_TAG);
-        }
-    }
-
-    @Override
-    public void onPause(){
-        super.onPause();
-        WriteModeOff();
-    }
-
-    @Override
-    public void onResume(){
-        super.onResume();
-        WriteModeOn();
-    }
-
-
-
-    /******************************************************************************
-     **********************************Enable Write********************************
-     ******************************************************************************/
-    private void WriteModeOn(){
-        writeMode = true;
-        mNfcAdapter.enableForegroundDispatch(this, pendingIntent, writeTagFilters, null);
-    }
-    /******************************************************************************
-     **********************************Disable Write*******************************
-     ******************************************************************************/
-    private void WriteModeOff(){
-        writeMode = false;
-        mNfcAdapter.disableForegroundDispatch(this);
-    }
-}
