--- app/src/main/java/enib/gala/ScannedBarcodeActivity.java
@@ -1,177 +1,0 @@
-package enib.gala;
-
-import android.Manifest;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.util.SparseArray;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
-import android.widget.TextView;
-
-import com.google.android.gms.vision.CameraSource;
-import com.google.android.gms.vision.Detector;
-import com.google.android.gms.vision.barcode.Barcode;
-import com.google.android.gms.vision.barcode.BarcodeDetector;
-import com.r0adkll.slidr.Slidr;
-import com.r0adkll.slidr.model.SlidrConfig;
-import com.r0adkll.slidr.model.SlidrListener;
-
-import java.io.IOException;
-
-public class ScannedBarcodeActivity extends AppCompatActivity {
-
-
-    SurfaceView surfaceView;
-    TextView txtBarcodeValue;
-    private CameraSource cameraSource;
-    private static final int REQUEST_CAMERA_PERMISSION = 201;
-    String intentData = "";
-    boolean isEmail = false;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_scanned_barcode);
-        SlidrConfig config = new SlidrConfig.Builder()
-                                .listener(new SlidrListener(){
-                                    @Override
-                                    public void onSlideStateChanged(int state) {
-
-                                    }
-
-                                    @Override
-                                    public void onSlideChange(float percent) {
-
-                                    }
-
-                                    @Override
-                                    public void onSlideOpened() {
-
-                                    }
-
-                                    @Override
-                                    public void onSlideClosed() {
-                                        beforeEndActivity();
-                                    }
-                                })
-                                .build();
-
-        Slidr.attach(this, config);
-
-        initViews();
-    }
-
-    private void beforeEndActivity()
-    {
-        Intent data = new Intent();
-        data.setData(Uri.parse(txtBarcodeValue.getText().toString()));
-        setResult(RESULT_OK, data);
-    }
-
-    private void initViews() {
-        txtBarcodeValue = findViewById(R.id.txtBarcodeValue);
-        surfaceView = findViewById(R.id.surfaceView);
-    }
-
-    private void initialiseDetectorsAndSources() {
-
-//        Toast.makeText(getApplicationContext(), "Barcode scanner started", Toast.LENGTH_SHORT).show();
-
-        BarcodeDetector barcodeDetector = new BarcodeDetector.Builder(this)
-                .setBarcodeFormats(Barcode.ALL_FORMATS)
-                .build();
-
-        cameraSource = new CameraSource.Builder(this, barcodeDetector)
-                .setRequestedPreviewSize(1920, 1080)
-                .setAutoFocusEnabled(true) //you should add this feature
-                .build();
-
-        surfaceView.getHolder().addCallback(new SurfaceHolder.Callback() {
-            @Override
-            public void surfaceCreated(SurfaceHolder holder) {
-                try {
-                    if (ActivityCompat.checkSelfPermission(ScannedBarcodeActivity.this, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED) {
-                        cameraSource.start(surfaceView.getHolder());
-                    } else {
-                        ActivityCompat.requestPermissions(ScannedBarcodeActivity.this, new
-                                String[]{Manifest.permission.CAMERA}, REQUEST_CAMERA_PERMISSION);
-                    }
-
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-
-
-            }
-
-            @Override
-            public void surfaceChanged(SurfaceHolder holder, int format, int width, int height) {
-            }
-
-            @Override
-            public void surfaceDestroyed(SurfaceHolder holder) {
-                cameraSource.stop();
-            }
-        });
-
-
-        barcodeDetector.setProcessor(new Detector.Processor<Barcode>() {
-            @Override
-            public void release() {
-//                Toast.makeText(getApplicationContext(), "To prevent memory leaks barcode scanner has been stopped", Toast.LENGTH_SHORT).show();
-            }
-
-            @Override
-            public void receiveDetections(Detector.Detections<Barcode> detections) {
-                final SparseArray<Barcode> barcodes = detections.getDetectedItems();
-                if (barcodes.size() != 0) {
-
-
-                    txtBarcodeValue.post(new Runnable() {
-
-                        @Override
-                        public void run() {
-
-                            if (barcodes.valueAt(0).email != null) {
-                                txtBarcodeValue.removeCallbacks(null);
-                                intentData = barcodes.valueAt(0).email.address;
-                                txtBarcodeValue.setText(intentData);
-                                isEmail = true;
-                            } else {
-                                isEmail = false;
-                                intentData = barcodes.valueAt(0).displayValue;
-
-
-                            }
-                            txtBarcodeValue.setText(intentData);
-//                            Toast.makeText(getApplicationContext(),intentData, Toast.LENGTH_SHORT).show();
-                            beforeEndActivity();
-                            finish();
-                        }
-                    });
-
-                }
-            }
-        });
-    }
-
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        cameraSource.release();
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        initialiseDetectorsAndSources();
-
-
-    }
-}
