--- app/src/main/java/enib/gala/Store.java
@@ -1,16 +1,0 @@
-package enib.gala;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-import com.r0adkll.slidr.Slidr;
-
-public class Store extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_store);
-        Slidr.attach(this);
-    }
-}
