+++ app/src/main/res/menu/menu_show_ticket.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="enib.gala.ShowTicket">
+    tools:context="enib.app.ShowTicket">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
