+++ app/src/main/res/xml/pref_headers.xml
@@ -3,17 +3,17 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="enib.gala.Settings$GeneralPreferenceFragment"
+        android:fragment="cash.leeap.Settings$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="enib.gala.Settings$NotificationPreferenceFragment"
+        android:fragment="cash.leeap.Settings$NotificationPreferenceFragment"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/pref_header_notifications" />
 
     <header
-        android:fragment="enib.gala.Settings$DataSyncPreferenceFragment"
+        android:fragment="cash.leeap.Settings$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
